split config and type
